+++ reana_server/config.py
@@ -180,5 +180,8 @@ ADMIN_EMAIL = os.getenv("REANA_EMAIL_SENDER", "CHANGE_ME")
 
 # Workflow scheduler
 # ==================
-REANA_SCHEDULER_SECONDS_TO_WAIT_FOR_REANA_READY = 1
+REANA_SCHEDULER_SECONDS_TO_WAIT_FOR_REANA_READY = 60
 """How many seconds to wait between retries in case of REANA not ready to run more workflows."""
+
+REANA_SCHEDULER_RETRY_DELAY = 60000
+"""How many milliseconds for a workflow to be delayed once rescheduled."""